Use cmd /c and Fix Get-ChildItem (#60)
* Execute Windows commands via Powershell

* Remove :PS_HOME\appserv before psadmin call

* Removing OS Check for envsummary command

* Fix typo

* Fix Windows path separators for glob function

* Added -NoProfile to powershell call

* Working on web start/stop

* Add PS_PSADMIN_PATH for Windows Web Start

* Removing empty line when finding domains in powershell

* App/Prcs working well on Windows

* Remove debugging statement

* Cleaning up debugging statements

+++ lib/psadmin_plus.rb
@@ -2,6 +2,7 @@
 
 require 'rbconfig'
 require 'etc'
+require 'open3'
 
 def do_help
     puts "Usage: psa [command] <type> <domain>"
@@ -198,7 +199,8 @@ def do_stop(type, domain)
         when "linux"
             do_cmd("${PS_CFG_HOME?}/webserv/#{domain}/bin/stopPIA.sh")
         when "windows"
-            do_cmd("#{PS_PSADMIN_PATH}/psadmin -w shutdown! -d #{domain}".gsub('/','\\'))
+            # do_cmd("#{PS_PSADMIN_PATH}/psadmin -w shutdown! -d #{domain}".gsub('/','\\'))
+            do_cmd("cmd /c ${env:PS_CFG_HOME}/webserv/#{domain}/bin/stopPIA.cmd".gsub('/','\\'))
         end
     else
         puts "Invalid type, see psa help"
@@ -243,7 +245,7 @@ def do_purge(type, domain)
             do_cmd("rm -rf ${PS_CFG_HOME?}/webserv/#{domain}/applications/peoplesoft/PORTAL*/*/cache*/")
             puts "web cache purged"
         when "windows"
-            do_cmd("Remove-Item -recurse -force ${env:PS_CFG_HOME}/webserv/#{domain}/applications/peoplesoft/PORTAL*/*/cache*/".gsub('/','\\'))
+            do_cmd("Remove-Item $(Get-ChildItem ${env:PS_CFG_HOME}/webserv/#{domain}/applications/peoplesoft/PORTAL*/*/cache*/ | ?{ $_.PSIsContainer}) -recurse -force".gsub('/','\\'))
         end
     else
         puts "Invalid type, see psa help"
@@ -293,7 +295,12 @@ def do_poolrm(type,domain)
     if PS_POOL_MGMT == "on" then
         # Change this function to match your pool member removal process
         puts "Removing domain from load balanced pool..."
-        do_cmd("rm -f #{env('PS_CFG_HOME')}/webserv/#{domain}/applications/peoplesoft/PORTAL.war/#{ENV['PS_HEALTH_FILE']}")
+        case OS_CONST
+        when "linux"
+            do_cmd("rm -f #{env('PS_CFG_HOME')}/webserv/#{domain}/applications/peoplesoft/PORTAL.war/#{ENV['PS_HEALTH_FILE']}")
+        when "windows"
+            do_cmd("remove-item -force #{env('PS_CFG_HOME')}/webserv/#{domain}/applications/peoplesoft/PORTAL.war/#{ENV['PS_HEALTH_FILE']}")
+        end
         sleep(PS_HEALTH_TIME.to_i)
         puts "...domain removed from pool."
         puts ""
